First homeworks

+++ src/odevler/bolum01/Q11.java
@@ -4,7 +4,7 @@ public class Q11 {
     public static void main(String[] args) {
         int population = 312032486;
         double year_with_seconds = 365.0*24*60*60;
-        double birth = (double) year_with_seconds / 7;
+        double birth = year_with_seconds / 7;
         double death = year_with_seconds / 13;
         double immigrant = year_with_seconds / 45;
         System.out.println("In one year later population will increase "+ (int)(birth +immigrant - death) + " and new population will be " + (int)(population + (birth +immigrant - death)));
